+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Tool to record data from Senso
+- Command-line interface to update firmware on Senso via Ethernet
 
 ## [2.1.0] - 2018-09-18