+++ Changelog.md
@@ -11,6 +11,7 @@
 ### Removed
 
 - Remove centralized log sink
+- Windows builds are no longer code-signed as there is no official Windows support
 
 ## [2.1.0] - 2018-09-18